witness script
